Fix generic Wayland icon on Flatpak
On a desktop with KDE Plasma, it was being shown a generic Wayland icon for GPT4All (Flatpak) because it couldn't match the running window to the desktop entry. This was due to the missing StartupWMClass field in the .desktop file.

By adding "StartupWMClass=io.gpt4all.chat" the window-class mismatch was fixed and allowed KDE to display the correct icon.

Signed-off-by: Diogo Caveiro <50335724+diogocaveiro@users.noreply.github.com>